[Insight] Code should not be duplicated
 - Add duplicate methods in an abstract class Controller
