+++ src/AppBundle/Controller/AbstractController.php
@@ -1,0 +1,69 @@
+<?php
+
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+namespace AppBundle\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Doctrine\ORM\QueryBuilder;
+
+/**
+ * Description of AbstractController
+ *
+ * @author Laurent
+ */
+abstract class AbstractController extends Controller
+{
+    /**
+     * @param string $name  session name
+     * @param string $field field name
+     * @param string $type  sort type ("ASC"/"DESC")
+     */
+    protected function setOrder($name, $field, $type = 'ASC')
+    {
+        $this->getRequest()->getSession()->set('sort.' . $name, array('field' => $field, 'type' => $type));
+    }
+
+    /**
+     * @param  string $name
+     * @return array
+     */
+    protected function getOrder($name)
+    {
+        $session = $this->getRequest()->getSession();
+
+        return $session->has('sort.' . $name) ? $session->get('sort.' . $name) : null;
+    }
+
+    /**
+     * @param QueryBuilder $qb
+     * @param string       $name
+     */
+    protected function addQueryBuilderSort(QueryBuilder $qb, $name)
+    {
+        $alias = current($qb->getDQLPart('from'))->getAlias();
+        if (is_array($order = $this->getOrder($name))) {
+            $qb->orderBy($alias . '.' . $order['field'], $order['type']);
+        }
+    }
+
+    /**
+     * Create Delete form
+     *
+     * @param integer                       $id
+     * @param string                        $route
+     * @return \Symfony\Component\Form\Form
+     */
+    protected function createDeleteForm($id, $route)
+    {
+        return $this->createFormBuilder(null, array('attr' => array('id' => 'delete')))
+            ->setAction($this->generateUrl($route, array('id' => $id)))
+            ->setMethod('DELETE')
+            ->getForm()
+        ;
+    }
+}
