+++ src/AppBundle/Controller/Settings/Divers/ZoneStorageController.php
@@ -15,7 +15,7 @@
 namespace AppBundle\Controller\Settings\Divers;
 
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use AppBundle\Controller\AbstractController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
@@ -29,7 +29,7 @@ use AppBundle\Form\Type\ZoneStorageType;
  *
  * @Route("/admin/settings/divers/zonestorage")
  */
-class ZoneStorageController extends Controller
+class ZoneStorageController extends AbstractController
 {
     /**
      * Lists all ZoneStorage entities.
@@ -180,21 +180,5 @@ class ZoneStorageController extends Controller
         }
 
         return $this->redirectToRoute('admin_zonestorage');
-    }
-
-    /**
-     * Create Delete form
-     *
-     * @param integer                       $id
-     * @param string                        $route
-     * @return \Symfony\Component\Form\Form
-     */
-    protected function createDeleteForm($id, $route)
-    {
-        return $this->createFormBuilder(null, array('attr' => array('id' => 'delete')))
-            ->setAction($this->generateUrl($route, array('id' => $id)))
-            ->setMethod('DELETE')
-            ->getForm()
-        ;
     }
 }
